Remove explicit pull for docker images
Since dev-mon01 alchemy registry is v1 (legacy), the docker module
issues a warning  that ansible treats as error. This forced to add
a pull from shell before each docker run. This patch removes the
pull and handles instead the warning message in the docker module to
decide if it is an error or not.