Set labels of build information in the Docker Image
Add option to pass in build information to Docker image when it is
being built. E.x.
"docker build --build-arg git_commit_id=<GITCOMMITSHA>
                --build-arg git_remote_url=<GITREMOTEURL>
                --build-arg build_date=<BUILDDATE>
                --build-arg jenkins_build_id=<JENKINSBUILDID>
                --build-arg jenkins_build_number=<JENKINSBUILDNUMBER> ."
Also default variables are set just in case there is not a variable passed in.

This is a twofold change… first, what needs to be done is the Docker file needs
to define the labels that can be passed into it (done in the patch) second when
"docker build" is called "--build-arg git_commit_id=<GITCOMMITSHA>
                --build-arg git_remote_url=<GITREMOTEURL>
                --build-arg build_date=<BUILDDATE>
                --build-arg jenkins_build_id=<JENKINSBUILDID>
                --build-arg jenkins_build_number=<JENKINSBUILDNUMBER>"
should be passed in as well. (This will be a change to the Jenkins Job)

e.g. instead of:  "docker build ."   It should be:
"docker build --build-arg git_commit_id=<GITCOMMITSHA>
                --build-arg git_remote_url=<GITREMOTEURL>
                --build-arg build_date=<BUILDDATE>
                --build-arg jenkins_build_id=<JENKINSBUILDID>
                --build-arg jenkins_build_number=<JENKINSBUILDNUMBER> ."