Switched kubernetes branch to pick up CNI initialization debugging
Switched the branch of ibm-contribs/kubernetes that we build
from r1.3-annotlab-plugindir to r13_annotlab_dbgcni.
That adds debugging of the kubelet initialization code that
finds the CNI plugin.
That drops the branch that adds the argument that enables a CLI
argument adding a directory to the search path for CNI config files
because we put the config file in the default place.