Fixed some playbook bugs wrt Ansible variable types
In Ansible 1, `set_fact: some_var="{{ some expression }}"` sets the
variable to a string regardless of the Jinja2 type of that expression.
Every non-empty string's truth value is true.  Thus `set_fact:
some_var="{{ 1==0 }}"` leads to the behavior that `{% if some_var %}`
behaves as if `some_var` is true.

The workaround is that for every variable intended to hold a Boolean
value, each usage should be of the form `varname | bool`.