Various fixes, mostly to docs
Added a section to the Ansible doc describing how to cope when the
target machines can not pull from public sources.

Added more descriptions into `ansible/group_vars/all`.

Removed `network_kind` from `ansible/group_vars/all` because
that variable must be supplied on the command line.

Explained that in the tiny example VMs, the vagrant user is not
authorized to use docker so `sudo` must be used.

Removed unneeded `network_kind` setting from the invocation
of the `env-basics.yml` playbook in the tiny example Vagrantfile.

Updated the invocation of the shard playbook in the tiny example
Vagrantfile to specify flannel rather than bridge as the networking
plugin to use.